+++ src/RequestConfig.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Shoxcie\BatchHttpClient;
 
 use Closure;
+use Symfony\Contracts\HttpClient\Exception\ExceptionInterface;
 use Symfony\Contracts\HttpClient\ResponseInterface;
-use Throwable;
 
 final class RequestConfig
 {
@@ -14,26 +14,26 @@ final class RequestConfig
     public string $url;
     /** @var array<string, mixed> */
     public array $options = [];
-    /** @var array<string, mixed>|Closure(int, Throwable): array<string, mixed> */
+    /** @var array<string, mixed>|Closure(string, int, ExceptionInterface|InvalidResponseException): array<string, mixed> */
     public $retryOptions = [];
-    public bool $throwOnError = true;
+    public bool $throwOnExhausted = true;
     public bool $decodeJson = true;
     public int $maxRetries = 0;
     public bool $retryOnTransportException = true;
-    /** @var null|Closure(string, mixed, ResponseInterface): mixed */
+    /** @var null|Closure(string, int, mixed, ResponseInterface): mixed */
     public ?Closure $parseResponse = null;
 
     /**
      * @param array<string, mixed> $options
-     * @param array<string, mixed>|Closure(int, Throwable): array<string, mixed> $retryOptions
-     * @param null|Closure(string, mixed, ResponseInterface): mixed $parseResponse
+     * @param array<string, mixed>|Closure(string, int, ExceptionInterface|InvalidResponseException): array<string, mixed> $retryOptions
+     * @param null|Closure(string, int, mixed, ResponseInterface): mixed $parseResponse
      */
     public function __construct(
         string $method,
         string $url,
         array $options = [],
         $retryOptions = [],
-        bool $throwOnError = true,
+        bool $throwOnExhausted = true,
         bool $decodeJson = true,
         int $maxRetries = 0,
         bool $retryOnTransportException = true,
@@ -43,7 +43,7 @@ final class RequestConfig
         $this->url = $url;
         $this->options = $options;
         $this->retryOptions = $retryOptions;
-        $this->throwOnError = $throwOnError;
+        $this->throwOnExhausted = $throwOnExhausted;
         $this->decodeJson = $decodeJson;
         $this->maxRetries = $maxRetries;
         $this->retryOnTransportException = $retryOnTransportException;